[GEWISDB] Move any `Member` association to `SubDecision` itself
Changes from GEWISDB upstream, see GEWIS/gewisdb#363. Necessary for
GDPR export functionality.
