+++ module/Decision/src/Model/SubDecision.php
@@ -137,6 +137,20 @@ abstract class SubDecision
     protected string $content;
 
     /**
+     * The member involved in this sub-decision.
+     *
+     * Not all sub-decisions require this, as such it is nullable. However, sub-decisions that need the guarantee that
+     * this is not null or need to specify an inverse side can do so using an association override.
+     */
+    #[ManyToOne(targetEntity: Member::class)]
+    #[JoinColumn(
+        name: 'lidnr',
+        referencedColumnName: 'lidnr',
+        nullable: true,
+    )]
+    protected ?Member $member = null;
+
+    /**
      * Get the decision.
      */
     public function getDecision(): Decision
@@ -203,6 +217,22 @@ abstract class SubDecision
     public function setNumber(int $number): void
     {
         $this->number = $number;
+    }
+
+    /**
+     * Get the member.
+     */
+    public function getMember(): ?Member
+    {
+        return $this->member;
+    }
+
+    /**
+     * Set the member.
+     */
+    public function setMember(Member $member): void
+    {
+        $this->member = $member;
     }
 
     /**
