+++ module/Decision/src/Model/SubDecision/Board/Installation.php
@@ -8,16 +8,27 @@ use DateTime;
 use Decision\Model\BoardMember;
 use Decision\Model\Member;
 use Decision\Model\SubDecision;
+use Doctrine\ORM\Mapping\AssociationOverride;
+use Doctrine\ORM\Mapping\AssociationOverrides;
 use Doctrine\ORM\Mapping\Column;
 use Doctrine\ORM\Mapping\Entity;
 use Doctrine\ORM\Mapping\JoinColumn;
-use Doctrine\ORM\Mapping\ManyToOne;
 use Doctrine\ORM\Mapping\OneToOne;
 
 /**
  * Installation as board member.
  */
 #[Entity]
+#[AssociationOverrides([
+    new AssociationOverride(
+        name: 'member',
+        joinColumns: new JoinColumn(
+            name: 'lidnr',
+            referencedColumnName: 'lidnr',
+            nullable: false,
+        ),
+    ),
+])]
 class Installation extends SubDecision
 {
     /**
@@ -25,21 +36,6 @@ class Installation extends SubDecision
      */
     #[Column(type: 'string')]
     protected string $function;
-
-    /**
-     * Member.
-     *
-     * Note that only members that are older than 18 years can be board members.
-     * Also, honorary, external and extraordinary members cannot be board members.
-     * (See the Statuten, Art. 13 Lid 2.
-     */
-    // TODO: Inversed relation
-    #[ManyToOne(targetEntity: Member::class)]
-    #[JoinColumn(
-        name: 'lidnr',
-        referencedColumnName: 'lidnr',
-    )]
-    protected Member $member;
 
     /**
      * The date at which the installation is in effect.
@@ -92,18 +88,12 @@ class Installation extends SubDecision
 
     /**
      * Get the member.
+     *
+     * @psalm-suppress InvalidNullableReturnType
      */
     public function getMember(): Member
     {
         return $this->member;
-    }
-
-    /**
-     * Set the member.
-     */
-    public function setMember(Member $member): void
-    {
-        $this->member = $member;
     }
 
     /**
