+++ module/Decision/src/Model/SubDecision/Key/Granting.php
@@ -6,28 +6,14 @@ namespace Decision\Model\SubDecision\Key;
 
 use DateTime;
 use Decision\Model\Keyholder;
-use Decision\Model\Member;
 use Decision\Model\SubDecision;
 use Doctrine\ORM\Mapping\Column;
 use Doctrine\ORM\Mapping\Entity;
-use Doctrine\ORM\Mapping\JoinColumn;
-use Doctrine\ORM\Mapping\ManyToOne;
 use Doctrine\ORM\Mapping\OneToOne;
 
 #[Entity]
 class Granting extends SubDecision
 {
-    /**
-     * The member who is granted a keycode of GEWIS.
-     */
-    #[ManyToOne(targetEntity: Member::class)]
-    #[JoinColumn(
-        name: 'lidnr',
-        referencedColumnName: 'lidnr',
-        nullable: true,
-    )]
-    protected ?Member $grantee = null;
-
     /**
      * Till when the keycode is granted.
      */
@@ -51,22 +37,6 @@ class Granting extends SubDecision
         mappedBy: 'grantingDec',
     )]
     protected Keyholder $keyholder;
-
-    /**
-     * Get the grantee.
-     */
-    public function getGrantee(): ?Member
-    {
-        return $this->grantee;
-    }
-
-    /**
-     * Set the grantee.
-     */
-    public function setGrantee(Member $grantee): void
-    {
-        $this->grantee = $grantee;
-    }
 
     /**
      * Get the date.
